requirements.txt
# Flask and web framework dependencies
Flask==3.0.0
Werkzeug==3.0.1

# Google Generative AI (Gemini)
google-generativeai==0.3.2

# HTTP requests
requests==2.31.0

# Environment variable management
python-dotenv==1.0.0

# JSON handling (usually included with Python)
# json - built-in module

# Logging (built-in with Python)
# logging - built-in module

# Time utilities (built-in with Python)
# time - built-in module

# OS utilities (built-in with Python)
# os - built-in module
